# api-specs/openapi.yml
openapi: 3.0.3
info:
  title: "SecureVault API"
  description: "Zero-knowledge Password Manager API. Handles encrypted blobs only."
  version: "1.0.0"

servers:
  - url: "http://localhost:8080/"
    description: "Local development server"

paths:
  /api/v1/auth/register:
    post:
      summary: "Register a new user"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserAuth'
      responses:
        '201':
          description: "User created"
        '400':
          description: "Validation error"
  /api/v1/auth/login:
      post:
        summary: Login user and get JWT
        operationId: post-api-v1-auth-login
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserAuth'
        responses:
          '200':
            description: Login successful
            content:
              application/json:
                schema:
                  type: object
                  properties:
                    token:
                      type: string
          '401':
            description: Invalid credentials

  /api/v1/vault:
    get:
      summary: "Get all secrets"
      security:
        - bearerAuth: []
      responses:
        '200':
          description: "List of encrypted secrets"
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Secret'

    post:
        summary: "Store a new secret"
        security:
          - bearerAuth: []
        requestBody:
          required: true
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateSecretRequest'
        responses:
          '201':
            description: "Secret created"
          '400':
            description: "Validation error"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    UserAuth:
      type: object
      required: [username, password_hash]
      properties:
        username:
          type: string
        password_hash:
          type: string
          description: "Client-side derived hash of the master password"

    CreateSecretRequest:
      type: object
      additionalProperties: false
      required:
        - encrypted_site_name
        - site_name_iv
        - encrypted_site_username
        - site_username_iv
        - encrypted_password
        - iv
      properties:
        encrypted_site_name:
          type: string
        site_name_iv:
          type: string
        encrypted_site_username:
          type: string
        site_username_iv:
          type: string
        encrypted_password:
          type: string
        iv:
          type: string
        enc_version:
          type: string
          default: v1

    Secret:
      type: object
      required:
        - id
        - encrypted_password
        - iv
      properties:
        id:
          type: integer
        encrypted_site_name:
          type: string
        site_name_iv:
          type: string
        encrypted_site_username:
          type: string
        site_username_iv:
          type: string
        encrypted_password:
          type: string
        iv:
          type: string
        enc_version:
          type: string